Start basic example
The big problem I'm facing is that esbuild [does not currently have web worker support](https://github.com/evanw/esbuild/issues/312), which is more to say it doesn't do magic to make it work yet. Web workers getting code by importing it from an available file basically gives the same problem as CSS for web components, but even trickier, since the import, resolve, and load isn't as simple. There may be a future workaround [here](https://github.com/evanw/esbuild/issues/312#issuecomment-1149442717), but overall this may just need to branch off and un-web-worker-ify the live parser.

I think this is ok though, because, for example, when implementing replays, I won't expect everything to work immediately with SlippiLab. It's more that once enough features are built, it would be able to be re-integrated. It can be the same with the live viewer and SpectatorMode.