Fix animations disappearing on bridge reconnect
The reason this would happen is that it seems resources are not fetched if the source returns `undefined` (or `null` for that matter), as was happening for all error cases. Due to this and the fact that a disconnect and reconnect doesn't reload the whole page, the resource would still have the old character ID, and would not think the animation needed to be re-fetched when the stream was re-loaded, even though re-loading the stream resets the animations.